🤖 Merge PR #59371 feat: support tiny-async-pool v2 by @mlecoq
