+++ types/tiny-async-pool/v1/index.d.ts
@@ -1,4 +1,4 @@
-// Type definitions for tiny-async-pool 2.0
+// Type definitions for tiny-async-pool 1.0
 // Project: https://github.com/rxaviers/async-pool#readme
 // Definitions by: Karl-Philipp Wulfert <https://github.com/krlwlfrt>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
@@ -25,5 +25,5 @@ export = asyncPool;
 declare function asyncPool<IN, OUT>(
     poolLimit: number,
     array: ReadonlyArray<IN>,
-    iteratorFn: (generator: IN) => Promise<OUT>,
-): AsyncIterableIterator<OUT>;
+    iteratorFn: (generator: IN) => Promise<OUT>
+): Promise<OUT[]>;
