+++ types/tiny-async-pool/v1/tsconfig.json
@@ -1,16 +1,19 @@
 {
     "compilerOptions": {
         "module": "commonjs",
-        "lib": ["es2018"],
+        "lib": ["es6"],
         "noImplicitAny": true,
         "noImplicitThis": true,
         "strictNullChecks": true,
         "strictFunctionTypes": true,
-        "baseUrl": "../",
-        "typeRoots": ["../"],
+        "baseUrl": "../../",
+        "typeRoots": ["../../"],
         "types": [],
         "noEmit": true,
-        "forceConsistentCasingInFileNames": true
+        "forceConsistentCasingInFileNames": true,
+        "paths": {
+            "tiny-async-pool": ["tiny-async-pool/v1"]
+        }
     },
     "files": ["index.d.ts", "tiny-async-pool-tests.ts"]
 }
